divide-by-components
