+++ src/page/main/MainBasket/CardSupplementMainBasket/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CardSupplementBasket";
