+++ src/page/main/MainBasket/CardTeaMainBasket/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CardTeaMainBasket";
