--- src/page/main/MainBasket/HeaderUserBasket/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./userBasket";
