+++ src/page/main/MainTea/DescriptionMainTea/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DescriptionMainTea";
